pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stocks-analyser"
version = "0.1.0"
description = "Tickertape + yfinance based stock analysis and paper-trading agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.44.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "yfinance>=0.2.54",
    "requests>=2.32.0",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
